Upgrade headspace with some time tracking
This is kind of ugly, and there's much that could be dried up, but
basically... I can tag things and say they should always be present or
never be present by either bundleID or tag.

I can still use `setup()` to be a more custom arrangement of things.

I also dry'd up the time tracking code, though it's kind of ugly too.