Moving towards using hs.layout for autolayout
I was manually moving the windows around with a series of `hs.window`
commands, but I think that building up an array of `hs.layout` compliant
tables and applying them will be better.

It should also let me declare more complex window layouts in init.lua if
I so desire.