Add to weekly ritual: one sentence summary
For a while earlier this year, I was recording my overall impression of
the week and a simple "rating" in my bullet journal. I am tempted to go
back and "fill in" for all the weeks that I have missed, but I think
that would slant my impression of the time. I have to do it in the now.