Experiment: Single page Review
I've been doing a two screen review, but I discovered today that
holding `⌘` and clicking a link in Things opens it in a new window.
(This also works for Quick Search!)

This means that a lot of my Review tasks could open in a new window...
I'm not sure how this will work for weekly, because I still would like
the checklist, but I'm going to give it a whirl.