Add the ability to require an intention
I'm noticing that I enter some of my more free-form spaces just to
"escape" something stricter... and then I get lost forever.

I've added the option to require a description for a space. If you put a
description in the space, you won't get pinged, but otherwise it prompts
for an intention in this space.

`{ requires_note = true }`